+++ general-squid/schema.graphql
@@ -1,10 +1,10 @@
 type Block @entity {
   id: ID! @index
   height: Int! @index
-  hash: Bytes! @index
-  parentHash: Bytes!
-  stateRoot: Bytes!
-  extrinsicsRoot: Bytes!
+  hash: String! @index
+  parentHash: String!
+  stateRoot: String!
+  extrinsicsRoot: String!
   specId: String!
   specName: String!
   specVersion: Int! @index
@@ -49,7 +49,7 @@ type Extrinsic @entity {
   fee: BigInt
   success: Boolean!
   error: JSON
-  hash: Bytes! @index
+  hash: String! @index
   timestamp: DateTime! @index
   name: String!
   signer: String
